requirements.txt
annotated-types @ file:///home/conda/feedstock_root/build_artifacts/annotated-types_1733247046149/work
anyio @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_anyio_1754315087/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1754231422783/work/certifi
click @ file:///D:/bld/click_1747811308011/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
deprecation==2.1.0
distro==1.9.0
dnspython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_dnspython_1757292856/work
email-validator @ file:///home/conda/feedstock_root/build_artifacts/email-validator-meta_1756221362582/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
fastapi @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_fastapi-core_1758060755/work
fastapi-cli @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_fastapi-cli_1757494948/work
h11 @ file:///home/conda/feedstock_root/build_artifacts/h11_1745526374115/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_httpcore_1745602916/work
httptools @ file:///D:/bld/httptools_1756754337472/work
httpx @ file:///home/conda/feedstock_root/build_artifacts/httpx_1733663348460/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
jiter==0.11.0
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1754951200865/work
MarkupSafe @ file:///D:/bld/markupsafe_1733219610373/work
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1733255585584/work
openai==1.107.3
packaging==25.0
postgrest==1.1.1
pydantic @ file:///home/conda/feedstock_root/build_artifacts/pydantic_1757881615017/work
pydantic_core @ file:///D:/bld/bld/rattler-build_pydantic-core_1746625416/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dotenv_1750789290/work
python-multipart @ file:///home/conda/feedstock_root/build_artifacts/python-multipart_1734420773152/work
PyYAML @ file:///D:/bld/pyyaml_1737454681229/work
realtime==2.7.0
rich @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_rich_1753436991/work/dist
rich-toolkit @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_rich-toolkit_1756998936/work
shellingham @ file:///home/conda/feedstock_root/build_artifacts/shellingham_1733300899265/work
six==1.17.0
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1733244044561/work
starlette @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_starlette_1757860651/work
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
tqdm==4.67.1
typer==0.17.4
typer-slim==0.17.4
typing-inspection @ file:///home/conda/feedstock_root/build_artifacts/typing-inspection_1747870647094/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
uvicorn @ file:///D:/bld/uvicorn_1751201629880/work
watchfiles @ file:///D:/bld/watchfiles_1757403445242/work
websockets @ file:///D:/bld/bld/rattler-build_websockets_1756476361/work/dist
